05_deep_learning/tf_regression_hours/requirements.txt
numpy
pandas
tensorflow-macos
tensorflow-metal